New batch_download.py (in dev)
