+++ src/encoded/static/components/item-pages/CaseView/InterpretationTab.js
@@ -22,14 +22,38 @@ export const InterpretationTab = React.memo(function InterpretationTab (props) {
                     Interpretation
                 </h1>
                 <div>
-                    <button type="button" className="btn btn-primary" disabled>
-                        Export all selections as <span className="text-600">TSV spreadsheet</span>
-                    </button>
+                    <ExportSpreadsheetButton {...{ variantSampleListItem }} />
                 </div>
             </div>
             <div>
                 <VariantSampleSelectionList {...{ isLoadingVariantSampleListItem, variantSampleListItem, schemas, context }} />
             </div>
         </React.Fragment>
+    );
+});
+
+
+const ExportSpreadsheetButton = React.memo(function ExportSpreadsheetButton({ variantSampleListItem }) {
+    const { "@id": atId, variant_samples: vsObjects = [] } = variantSampleListItem || {};
+    const baseHref = "/variant-sample-list-spreadsheet/?variant_sample_list_id=" + encodeURIComponent(atId);
+    return (
+        <DropdownButton variant="primary" disabled={vsObjects.length === 0} title="Export as...">
+            {/* Testing -- }
+            <form method="POST" action={baseHref + "&file_format=tsv"} className="d-inline-block">
+                <button type="submit" className="dropdown-item" role="button" download>
+                    <span className="text-600">TSV</span> spreadsheet
+                </button>
+            </form>
+            */}
+            <a href={baseHref + "&file_format=tsv"} target="_blank" rel="noopener noreferrer" className="dropdown-item" role="button" download>
+                <span className="text-600">TSV</span> spreadsheet
+            </a>
+            <a href={baseHref + "&file_format=csv"} target="_blank" rel="noopener noreferrer" className="dropdown-item" role="button" download>
+                <span className="text-600">CSV</span> spreadsheet
+            </a>
+            <a href="#" className="dropdown-item disabled" target="_blank" rel="noopener noreferrer" role="button">
+                <span className="text-600">XLSX</span> spreadsheet
+            </a>
+        </DropdownButton>
     );
 });
