+++ src/encoded/util.py
@@ -319,7 +319,8 @@ CONTENT_TYPE_SPECIAL_CASES = {
     'application/x-www-form-urlencoded': [
         # Single legacy special case to allow us to POST to metadata TSV requests via form submission.
         # All other special case values should be added using register_path_content_type.
-        '/metadata/'
+        '/metadata/',
+        '/variant-sample-list-spreadsheet/'
     ]
 }
